pyproject.toml
[project]
name = "emailvalidator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dnspython>=2.7.0",
    "pydantic>=2.9.2",
    "requests>=2.32.3",
]
